nfsd: ensure we use ctime_peek to grab the inode->i_ctime
If getattr fails, then nfsd can end up scraping the time values directly
out of the inode for pre and post-op attrs. This may or may not be the
right thing to do, but for now make it at least use ctime_peek in this
situation to ensure that the QUERIED flag is masked.

Signed-off-by: Jeff Layton <jlayton@kernel.org>
Commit: J. Bruce Fields <bfields@redhat.com>
Signed-off-by: Amir Goldstein <amir73il@gmail.com>
Signed-off-by: J. Bruce Fields <bfields@redhat.com>
Message-Id: <20230518114742.128950-5-jlayton@kernel.org>
Signed-off-by: Christian Brauner <brauner@kernel.org>